Inline simplified internal route list helper class
