--- packages/framework/tests/Feature/RouteListTest.php
@@ -1,82 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Hyde\Framework\Testing\Feature;
-
-use Hyde\Hyde;
-use Hyde\Testing\TestCase;
-use Hyde\Pages\InMemoryPage;
-use Hyde\Support\Models\Route;
-use Hyde\Support\Internal\RouteList;
-
-/**
- * @covers \Hyde\Support\Internal\RouteList
- * @covers \Hyde\Support\Internal\RouteListItem
- */
-class RouteListTest extends TestCase
-{
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        [$page404, $pageIndex] = Hyde::routes()->values();
-
-        // Run without files to not use the href links
-        $this->withoutDefaultPages();
-
-        Hyde::routes()->put('404', $page404);
-        Hyde::routes()->put('index', $pageIndex);
-    }
-
-    protected function tearDown(): void
-    {
-        $this->restoreDefaultPages();
-
-        parent::tearDown();
-    }
-
-    public function testRouteList()
-    {
-        $this->assertSame([
-            [
-                'page_type' => 'BladePage',
-                'source_file' => '_pages/404.blade.php',
-                'output_file' => '_site/404.html',
-                'route_key' => '404',
-            ],
-            [
-                'page_type' => 'BladePage',
-                'source_file' => '_pages/index.blade.php',
-                'output_file' => '_site/index.html',
-                'route_key' => 'index',
-            ],
-        ], (new RouteList())->rows());
-    }
-
-    public function testHeaders()
-    {
-        $this->assertSame([
-            'Page Type',
-            'Source File',
-            'Output File',
-            'Route Key',
-        ], (new RouteList())->header());
-    }
-
-    public function testWithDynamicPages()
-    {
-        Hyde::routes()->forget('404');
-        Hyde::routes()->forget('index');
-        Hyde::routes()->put('foo', new Route(new InMemoryPage('foo')));
-
-        $this->assertSame([
-            [
-                'page_type' => 'InMemoryPage',
-                'source_file' => '<fg=gray>none</>',
-                'output_file' => '_site/foo.html',
-                'route_key' => 'foo',
-            ],
-        ], (new RouteList())->rows());
-    }
-}
